+++ lib/Braintree/Test/Nonces.php
@@ -23,10 +23,15 @@ class Braintree_Test_Nonces
    public static $consumed = "fake-consumed-nonce";
    public static $paypalOneTimePayment = "fake-paypal-one-time-nonce";
    public static $paypalFuturePayment = "fake-paypal-future-nonce";
+   public static $paypalBillingAgreement = "fake-paypal-billing-agreement-nonce";
    public static $applePayVisa = "fake-apple-pay-visa-nonce";
    public static $applePayMasterCard = "fake-apple-pay-visa-nonce";
    public static $applePayAmEx = "fake-apple-pay-amex-nonce";
    public static $androidPay = "fake-android-pay-nonce";
+   public static $androidPayDiscover = "fake-android-pay-discover-nonce";
+   public static $androidPayVisa = "fake-android-pay-visa-nonce";
+   public static $androidPayMasterCard = "fake-android-pay-mastercard-nonce";
+   public static $androidPayAmEx = "fake-android-pay-amex-nonce";
    public static $abstractTransactable = "fake-abstract-transactable-nonce";
    public static $europe = "fake-europe-bank-account-nonce";
    public static $coinbase = "fake-coinbase-nonce";