+++ lib/Braintree/Transaction/AndroidPayCardDetails.php
@@ -23,6 +23,7 @@
  * @property-read string $imageUrl
  * @property-read string $sourceCardLast4
  * @property-read string $sourceCardType
+ * @property-read string $sourceDescription
  * @property-read string $token
  * @property-read string $virtualCardLast4
  * @property-read string $virtualCardType